+++ packages/frontend-near-wallet/README.md
@@ -1,0 +1,3 @@
+# NEAR Wallet Post Migration Landing Page
+
+
