+++ packages/frontend-near-wallet/src/modules/redirects/pages/RedirectPage.tsx
@@ -4,8 +4,9 @@ import { getMyNearWalletUrl } from '../../../utils/constants';
 
 const RedirectPage = () => {
     const location = useLocation()
+    // TODO: Modify logic here to redirect based on user's selection
     const destinationWalletBaseURL = getMyNearWalletUrl();
-    const urlString = location.pathname + location.search
+    const urlString = location.pathname + location.search;
     let redirectURL = new URL(urlString, destinationWalletBaseURL);
 
     window.location.href = redirectURL.toString();
